+++ runtime/src/main/java/com/kotcrab/vis/runtime/system/inflater/MusicInflater.java
@@ -58,7 +58,6 @@ public class MusicInflater extends InflaterSystem {
 		protoVisMusic.fill(musicComponent);
 		if (musicComponent.playOnStart) musicComponent.music.play();
 
-		if (configuration.removeAssetsComponentAfterInflating) assetCm.remove(entityId);
 		protoCm.remove(entityId);
 	}
 }